+++ tests/testthat/test-compute-boxplot.r
@@ -10,7 +10,7 @@ test_that("Zero-row inputs", {
   expect_identical(res$outliers, list())
 
   # Grouped
-  res <- mtcars %>% group_by(cyl) %>% filter(FALSE) %>% compute_boxplot(~mpg)
+  res <- mtcars %>% group_by(cyl) %>% dplyr::filter(FALSE) %>% compute_boxplot(~mpg)
   expect_equal(nrow(res), 0)
   expect_true(setequal(
     names(res),