+++ tests/testthat/test-compute-count.r
@@ -53,7 +53,7 @@ test_that("Zero-row inputs", {
   expect_true(setequal(names(res), c("count_", "x_")))
 
   # Grouped
-  res <- mtcars %>% group_by(cyl) %>% filter(FALSE) %>% compute_count(~cyl)
+  res <- mtcars %>% group_by(cyl) %>% dplyr::filter(FALSE) %>% compute_count(~cyl)
   expect_equal(nrow(res), 0)
   expect_true(setequal(names(res), c("cyl", "count_", "x_")))
 })