+++ tests/testthat/test-compute-density.r
@@ -20,7 +20,7 @@ test_that("Zero-row inputs", {
   expect_true(setequal(names(res), c("pred_", "resp_")))
 
   # Grouped
-  res <- mtcars %>% group_by(cyl) %>% filter(FALSE) %>% compute_density(~mpg)
+  res <- mtcars %>% group_by(cyl) %>% dplyr::filter(FALSE) %>% compute_density(~mpg)
   expect_equal(nrow(res), 0)
   expect_true(setequal(names(res), c("cyl", "pred_", "resp_")))
 })