+++ tests/testthat/test-compute-tabulate.r
@@ -6,7 +6,7 @@ test_that("Zero-row inputs", {
   expect_true(setequal(names(res), c("count_", "x_")))
 
   # Grouped
-  res <- mtcars %>% group_by(cyl) %>% filter(FALSE) %>% compute_tabulate(~factor(cyl))
+  res <- mtcars %>% group_by(cyl) %>% dplyr::filter(FALSE) %>% compute_tabulate(~factor(cyl))
   expect_equal(nrow(res), 0)
   expect_true(setequal(names(res), c("cyl", "count_", "x_")))
 })