+++ NEWS.md
@@ -1,5 +1,9 @@
 # ggvis 0.3.0.9xxx
 
+* `filter()` is no longer imported and re-exported from dplyr. This
+  means that to use `filter()` with ggvis object you'll need to 
+  make sure to load dplyr first.
+
 * `compute_smooth()` supports more complex formulas. (#209)
 
 * Boxplots are now supported, with `layer_boxplots()` and `compute_boxplot()`.